Refactor file imports in layout and subjects pages
